restructuring my code files for maintainability and easier to navigate
